First draft of strategy computations
The computation of values/strategies was moved to new functions
`energy_levels._compute_{objective}`. These are called from
`energy_levels.get_{objective}` which allows to keep the current
interface. Strategies are stored in new dict, where a strategy
for OBEJECTIVE is in `energy_levels.strategy[OBJECTIVE]`.

We can implement the values in similar fashion later on.