ADD: *Counter strategies* and strategy interface.
Interface for strategies defines
 1. how to ask a strategy for the next action to be played, and
 2. how to update strategy about the outcome of the picked action

The interface enforces that a strategy is always informed about
the evolution of the play.

`CounterStrategy` is an implementation of this interface. It
maintains the counter strategies proposed in our CAV paper and
relies on `CounterSelectors` for choosing the actions.

`CounterSelectors` are enriched equivalent of the current notion
of "strategy". Finally, `SelectionRule` is a dict that translates
itself into a interval-based mapping.