DBAWrapper as convenient interfact to Spot's DBAs.
`fimdp.products.DBAWrapper` is initialized using a Spot's
representation of a determinstic automaton and a list of names
of atomic propositions.

The functions `succ` and `edge` take `state` and `label` as
arguments and return the corresponding successor or edge. Note
that `labels` is a sequence of indices into the list of AP names
given at creation of the DBAWrapper object.