ProductSelector & ProductSelectorWrapper
These two selectors are suited for MDPs that need some product for
analysis. The selection rules are built using states and actions
of the product, but the selection itself (`selection.select_action`)
should use states of the original MDP (before building product) and
returns also actions of the original MDP and not of product.

The ..Wrapper class is basically an interface that translates the
values of the selector built for the product into the actions of
the original MDP. It stores the values in the same way as a
CounterSelector build for the product. After product is deleted,
the selector cannot be used for selection.

ProductSelector has a unique method of data storage and saves the
actions of the original MDP directly at the time when `update()`
is called. Therefore, `select_action` can be called after the
product MDP is destroyed.

CounterSelectors now have `copy_values_from(other_selector)`.