ProductSelectors for labelled CMDPs + DBACounterStrategy
LCMDP learned selector_for_dba which constructs product with given
DBA, analyze the product for Büchi objective using ProductSelector.
The product selector is returned and the product is destroyed.

The ProductSelectors can be used in DBACounterStrategies that have
counter and a DBA in memory. They work seemlessly on top of labelled
CMDPs.